feat: improve ios touch gesture
